# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "LiventCord",
    "version": "1.0"
  },
  "paths": {
    "/api/guilds/{guildId}/channels": {
      "get": {
        "tags": [
          "Channel"
        ],
        "parameters": [
          {
            "name": "guildId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/{guildId}/channels/{channelId}": {
      "delete": {
        "tags": [
          "Channel"
        ],
        "parameters": [
          {
            "name": "guildId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channelId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/{guildId}/channels": {
      "post": {
        "tags": [
          "Channel"
        ],
        "parameters": [
          {
            "name": "guildId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChannelRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChannelRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChannelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/channels/{guildId}/{channelId}": {
      "get": {
        "tags": [
          "LiventCord"
        ],
        "parameters": [
          {
            "name": "guildId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channelId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/channels/{friendId}": {
      "get": {
        "tags": [
          "LiventCord"
        ],
        "parameters": [
          {
            "name": "friendId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/guilds/{guildId}/channels/{channelId}/messages": {
      "post": {
        "tags": [
          "Message"
        ],
        "parameters": [
          {
            "name": "guildId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channelId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewMessageRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/NewMessageRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/NewMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Message"
        ],
        "parameters": [
          {
            "name": "GuildId",
            "in": "query",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ChannelId",
            "in": "query",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "guildId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channelId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/guilds/{guildId}/channels/{channelId}/messages/edit": {
      "post": {
        "tags": [
          "Message"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "header",
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "guildId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channelId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditMessageRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EditMessageRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EditMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/search": {
      "get": {
        "tags": [
          "Message"
        ],
        "parameters": [
          {
            "name": "guildId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Channel": {
        "type": "object",
        "properties": {
          "channelId": {
            "type": "string",
            "nullable": true
          },
          "channelName": {
            "type": "string",
            "nullable": true
          },
          "channelDescription": {
            "type": "string",
            "nullable": true
          },
          "isTextChannel": {
            "type": "boolean"
          },
          "lastReadDateTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "guildId": {
            "type": "string",
            "nullable": true
          },
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "guild": {
            "$ref": "#/components/schemas/Guild"
          },
          "userChannels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserChannel"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateChannelRequest": {
        "required": [
          "channelName",
          "isTextChannel"
        ],
        "type": "object",
        "properties": {
          "channelName": {
            "minLength": 1,
            "type": "string"
          },
          "isTextChannel": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "EditMessageRequest": {
        "required": [
          "channelId",
          "content",
          "guildId",
          "messageId"
        ],
        "type": "object",
        "properties": {
          "guildId": {
            "minLength": 1,
            "type": "string"
          },
          "messageId": {
            "minLength": 1,
            "type": "string"
          },
          "channelId": {
            "minLength": 1,
            "type": "string"
          },
          "content": {
            "maxLength": 2000,
            "minLength": 0,
            "type": "string"
          },
          "attachmentUrls": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Guild": {
        "type": "object",
        "properties": {
          "guildId": {
            "type": "string",
            "nullable": true
          },
          "ownerId": {
            "type": "string",
            "nullable": true
          },
          "guildName": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "rootChannel": {
            "type": "string",
            "nullable": true
          },
          "region": {
            "type": "string",
            "nullable": true
          },
          "settings": {
            "type": "string",
            "nullable": true
          },
          "isGuildUploadedImg": {
            "type": "boolean"
          },
          "guildMembers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GuildUser"
            },
            "nullable": true
          },
          "channels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Channel"
            },
            "nullable": true
          },
          "userIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "readOnly": true
          },
          "guildPermissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GuildPermissions"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GuildPermissions": {
        "type": "object",
        "properties": {
          "guildId": {
            "type": "string",
            "nullable": true
          },
          "guild": {
            "$ref": "#/components/schemas/Guild"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "readMessages": {
            "type": "integer",
            "format": "int32"
          },
          "sendMessages": {
            "type": "integer",
            "format": "int32"
          },
          "manageRoles": {
            "type": "integer",
            "format": "int32"
          },
          "kickMembers": {
            "type": "integer",
            "format": "int32"
          },
          "banMembers": {
            "type": "integer",
            "format": "int32"
          },
          "manageChannels": {
            "type": "integer",
            "format": "int32"
          },
          "mentionEveryone": {
            "type": "integer",
            "format": "int32"
          },
          "addReaction": {
            "type": "integer",
            "format": "int32"
          },
          "isAdmin": {
            "type": "integer",
            "format": "int32"
          },
          "canInvite": {
            "type": "integer",
            "format": "int32"
          },
          "permissions": {
            "$ref": "#/components/schemas/PermissionFlags"
          }
        },
        "additionalProperties": false
      },
      "GuildUser": {
        "type": "object",
        "properties": {
          "guildId": {
            "type": "string",
            "nullable": true
          },
          "memberId": {
            "type": "string",
            "nullable": true
          },
          "guild": {
            "$ref": "#/components/schemas/Guild"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "additionalProperties": false
      },
      "Message": {
        "type": "object",
        "properties": {
          "messageId": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "channelId": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "lastEdited": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "attachmentUrls": {
            "type": "string",
            "nullable": true
          },
          "replyToId": {
            "type": "string",
            "nullable": true
          },
          "reactionEmojisIds": {
            "type": "string",
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "channel": {
            "$ref": "#/components/schemas/Channel"
          }
        },
        "additionalProperties": false
      },
      "NewMessageRequest": {
        "required": [
          "channelId",
          "content",
          "guildId"
        ],
        "type": "object",
        "properties": {
          "guildId": {
            "minLength": 1,
            "type": "string"
          },
          "channelId": {
            "minLength": 1,
            "type": "string"
          },
          "content": {
            "maxLength": 2000,
            "minLength": 0,
            "type": "string"
          },
          "attachmentUrls": {
            "type": "string",
            "nullable": true
          },
          "replyToId": {
            "type": "string",
            "nullable": true
          },
          "reactionEmojisIds": {
            "type": "string",
            "nullable": true
          },
          "lastEdited": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PermissionFlags": {
        "enum": [
          0,
          1,
          2,
          4,
          8,
          16,
          32,
          64,
          128,
          256,
          512,
          1023
        ],
        "type": "integer",
        "format": "int32"
      },
      "User": {
        "required": [
          "bot",
          "discriminator",
          "email",
          "nickname",
          "password",
          "status"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "nullable": true
          },
          "guildUsers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GuildUser"
            },
            "nullable": true
          },
          "email": {
            "maxLength": 128,
            "minLength": 0,
            "type": "string"
          },
          "discriminator": {
            "maxLength": 4,
            "minLength": 0,
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 0,
            "type": "string"
          },
          "nickname": {
            "maxLength": 32,
            "minLength": 0,
            "type": "string"
          },
          "bot": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "maxLength": 128,
            "minLength": 0,
            "type": "string"
          },
          "description": {
            "maxLength": 256,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastLogin": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "verified": {
            "type": "integer",
            "format": "int32"
          },
          "location": {
            "maxLength": 256,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "language": {
            "maxLength": 10,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "socialMediaLinks": {
            "maxLength": 512,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "userChannels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserChannel"
            },
            "nullable": true
          },
          "guildPermissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GuildPermissions"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserChannel": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "nullable": true
          },
          "channelId": {
            "type": "string",
            "nullable": true
          },
          "lastReadDatetime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "channel": {
            "$ref": "#/components/schemas/Channel"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "additionalProperties": false
      }
    }
  }
}